+++ CHANGELOG.md
@@ -2,4 +2,6 @@
 
 * 0.0.1  initial release.
 * 0.0.2  Fixed issues.
-* 0.0.3  Added context extensions and fixed bugs.
+* 0.0.3  Added context extensions
+and fixed bugs.
+* 0.0.4  Fixed initial darkmode flag issue when current theme is dark mode .